MOdified program to hide input form web link.
