+++ onprc_ehr/src/org/labkey/onprc_ehr/dataentry/NecropsyRequestForm.java
@@ -9,7 +9,11 @@ import org.labkey.api.ehr.dataentry.NonStoreFormSection;
 import org.labkey.api.ehr.dataentry.RequestForm;
 import org.labkey.api.module.Module;
 import org.labkey.api.module.ModuleLoader;
+import org.labkey.api.security.Group;
+import org.labkey.api.security.GroupManager;
+import org.labkey.api.security.permissions.AdminPermission;
 import org.labkey.api.view.template.ClientDependency;
+import org.labkey.security.xml.GroupEnumType;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -60,6 +64,18 @@ public class NecropsyRequestForm extends RequestForm
         setStoreCollectionClass("onprc_ehr.data.sources.PathTissueRequestStoreCollection");
         addClientDependency(ClientDependency.supplierFromPath("onprc_ehr/data/sources/PathTissueRequestStoreCollection.js"));
    }
+
+    //Added: 8-7-2024 R.Blasa
+    @Override
+    public boolean isVisible()
+    {
+        Group g = GroupManager.getGroup(getCtx().getContainer(), "Death Entry", GroupEnumType.SITE);
+        if (g != null && getCtx().getUser().isInGroup(g.getUserId()) && !getCtx().getContainer().hasPermission(getCtx().getUser(), AdminPermission.class))
+        {
+            return false;
+        }
+        return super.isVisible();
+    }
     @Override
     public JSONObject toJSON()
     {
